Separate out the core service to be able to use from search
