+++ lib/Exceptions/UnsupportedFiletypeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace OCA\Metadata\Exceptions;
+
+class UnsupportedFiletypeException extends \Exception {
+
+}
